Permit single DDL commands in transaction blocks
This change uses the new XactOperations bit field to permit single DDL
commands within transaction blocks, functions, and other multi-command
transactional contexts. If a second DDL command or DML command is sub-
sequently issued, an error will be raised. Similarly, an error will be
raised if a DDL command is issued in a block which has processed DML
commands.